# swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Backend REST API",
    "version": "1.0"
  },
  "paths": {
...
    "/api/items/update/{id}": {
      "put": {
        "tags": ["Items"],
        "summary": "Updates an item, requires admin token",
        "responses": {
          "200": {
            "description": "Item updated."
          }
        },
        "parameters": [
          {
            "name": "ID",
            "required": true,
            "description": "ID of item to update",
            "schema": { "type": "string" },
            "in": "path"
          }
        ]
      }
    },
...
    "/api/items/get/{id}": {
      "get": {
        "tags": ["Items"],
        "summary": "Gets an item",
        "responses": {
          "200": {
            "description": "Returns a specific item"
          }
        },
        "parameters": [
          {
            "name": "ID",
            "required": true,
            "description": "ID of item to get",
            "schema": { "type": "string" },
            "in": "path"
          }
        ]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BasicAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}